Major Problems
Trying to make the server able to send CreateUserReply and LoginReply
messages.
Seemingly a lot harder than I thought it would be.

Since most of these changes happen in the ChatTwo_Protocol class which
is shared between the two app, I had to make a lot of changes in
ChatTwo_Client_Protocol.
I fear I broke more than I fixed this this.

ChatTwo_Protocol:
- Added support to have the UserID number as part of the messge header.
But this again open up for a lot of problems for messages where you have
no UserID number to send.